pyproject.toml
[project]
name = "github-copilot-thought-diary-app"
version = "0.1.0"
description = "A full-stack web application powered by AI for tracking thought diaries"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "flask>=3.0.0",
    "sqlalchemy>=2.0.0",
    "flask-sqlalchemy>=3.1.0",
    "alembic>=1.12.0",
    "bcrypt>=4.0.1",
    "email-validator>=2.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-jwt-extended>=4.7.1",
    "flask-limiter>=4.0.0",
    "flask-marshmallow>=1.3.0",
    "marshmallow-sqlalchemy>=1.4.2",
]
